Incorporate remaining upstream fixes from DraconPern/fmjpeg2koj
- Guard against empty compressed pixel items (compressedSize == 0)
  in decodeFrame to avoid undefined behaviour on corrupt data
- Fix buffer_size type from size_t to unsigned long in
  compressRenderedFrame for DCMTK API compatibility

+++ djcodece.cc
@@ -918,7 +918,7 @@ OFCondition DJPEG2KEncoderBase::compressRenderedFrame(
 
     // This is the buffer with the uncompressed pixel data
     Uint8 *buffer;
-    size_t buffer_size;
+    unsigned long buffer_size;
 
     const Uint32 framesize = dimage->getWidth() * dimage->getHeight();
     switch(dinter->getRepresentation())
